[P2] 巡检修复：flashPanicPerSec 类型修正 + 死代码清理 + 测试状态修复
- LevelConfigSO.flashPanicPerSec: int → float，支持小数慌乱增长率
- DevModeController: 删除未使用的 taskManager 字段，移除多余的 RoundToInt
- AutomatedGameTest: 修复 isTesting 从未设为 true 的 bug

+++ Pick-Light-From-Dark/Assets/Scripts/Game/Config/LevelConfigSO.cs
@@ -57,7 +57,7 @@ namespace Game.Config
         public Vector2 flashCheckDuration;
 
         [Header("手电筒")]
-        public int flashPanicPerSec;
+        public float flashPanicPerSec;
 
         [Header("初始卡牌")]
         public List<int> initialCards;
